Additional fix for atftpd upgrade
We need to remove the /etc/default/atftpd in preinst too:
the prerm of the previous version does not always fail depending on
systemd installation / configuration, so the prerm/failed-upgrade of the
new version is not executed. But the upgrade will still fail when the
postinst is executed.
In preinst we have to test if the /etc/default/atftpd exists: it may
have been removed by the prerm and the test fails if it doesn't exist.